engine: Increase precision of rx_rate and tx_rate
The precision of the rx_rate, tx_rate, rx_drop and tx_drop
of virtual and host network interfaces is increased.

This enables to detect little traffic on network interface
statistics.

Bug-Url: https://bugzilla.redhat.com/1459908

+++ backend/manager/modules/vdsbroker/src/main/java/org/ovirt/engine/core/vdsbroker/NetworkStatisticsBuilder.java
@@ -97,7 +97,7 @@ public class NetworkStatisticsBuilder {
             return null;
         }
 
-        long megabitDiff = BITS_IN_BYTE * byteDiff / BITS_IN_MEGABIT;
+        double megabitDiff = BITS_IN_BYTE * (double)byteDiff / BITS_IN_MEGABIT;
         double timeDiffInSeconds = currentTime - previousTime;
         double rateInMbps = megabitDiff / timeDiffInSeconds;
         return truncatePercentage(100 * rateInMbps / speed);
